requirements.txt
dbus-python
gi
